+++ src/com/intellij/idea/plugin/hybris/tools/ccm/SAPCCMDeploymentCommands.kt
@@ -1,0 +1,81 @@
+/*
+ * This file is part of "SAP Commerce Developers Toolset" plugin for IntelliJ IDEA.
+ * Copyright (C) 2019-2024 EPAM Systems <hybrisideaplugin@epam.com> and contributors
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU Lesser General Public License as
+ * published by the Free Software Foundation, either version 3 of the
+ * License, or (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.
+ * See the GNU Lesser General Public License for more details.
+ *
+ * You should have received a copy of the GNU Lesser General Public License
+ * along with this program. If not, see <http://www.gnu.org/licenses/>.
+ */
+
+package com.intellij.idea.plugin.hybris.tools.ccm
+
+import com.intellij.idea.plugin.hybris.settings.CCv2Subscription
+import com.intellij.idea.plugin.hybris.settings.components.ApplicationSettingsComponent
+import com.intellij.idea.plugin.hybris.tools.ccv2.dto.CCv2Deployment
+import com.intellij.idea.plugin.hybris.tools.ccv2.dto.CCv2DeploymentDatabaseUpdateModeEnum
+import com.intellij.idea.plugin.hybris.tools.ccv2.dto.CCv2DeploymentStrategyEnum
+import com.intellij.openapi.project.Project
+
+object SAPCCMDeploymentCommands {
+    private const val command = "deployment"
+    private val headers = listOf(
+        "CODE",
+        "CREATED BY",
+        "CREATED TIME",
+        "BUILD CODE",
+        "ENV. CODE",
+        "UPDATE MODE",
+        "STRATEGY",
+        "SCHEDULED TIME",
+        "DEPLOYED TIME",
+        "FAILED TIME",
+        "UNDEPLOYED TIME",
+        "STATUS",
+        "CANCELED BY",
+        "CANCELED TIME",
+        "CANCEL FINISHED TIME",
+        "CANCEL FAILED",
+    )
+
+    private val listCommand = object : AbstractSAPCCMListCommand<CCv2Deployment>("Deployments", command, headers) {}
+
+    fun list(
+        project: Project,
+        appSettings: ApplicationSettingsComponent,
+        subscriptions: Collection<CCv2Subscription>
+    ) = listCommand.list(project, appSettings, subscriptions) { row, columns -> mapToDTO(row, columns) }
+
+    private fun mapToDTO(row: String, columns: Map<String, Int>): CCv2Deployment {
+        return CCv2Deployment(
+            code = row.substring(0..<columns["CREATED BY"]!!).trim(),
+            createdBy = row.substring(columns["CREATED BY"]!!..<columns["CREATED TIME"]!!).trim(),
+            createdTime = row.substring(columns["CREATED TIME"]!!..<columns["BUILD CODE"]!!).trim(),
+            buildCode = row.substring(columns["BUILD CODE"]!!..<columns["ENV. CODE"]!!).trim(),
+            envCode = row.substring(columns["ENV. CODE"]!!..<columns["UPDATE MODE"]!!).trim(),
+            updateMode = CCv2DeploymentDatabaseUpdateModeEnum.tryValueOf(
+                row.substring(columns["UPDATE MODE"]!!..<columns["STRATEGY"]!!).trim()
+            ),
+            strategy = CCv2DeploymentStrategyEnum.tryValueOf(
+                row.substring(columns["STRATEGY"]!!..<columns["SCHEDULED TIME"]!!).trim()
+            ),
+            scheduledTime = row.substring(columns["SCHEDULED TIME"]!!..<columns["DEPLOYED TIME"]!!).trim(),
+            deployedTime = row.substring(columns["DEPLOYED TIME"]!!..<columns["FAILED TIME"]!!).trim(),
+            failedTime = row.substring(columns["FAILED TIME"]!!..<columns["UNDEPLOYED TIME"]!!).trim(),
+            undeployedTime = row.substring(columns["UNDEPLOYED TIME"]!!..<columns["STATUS"]!!).trim(),
+            status = row.substring(columns["STATUS"]!!..<columns["CANCELED BY"]!!).trim(),
+            cancelledBy = row.substring(columns["CANCELED BY"]!!..<columns["CANCELED TIME"]!!).trim(),
+            cancelledTime = row.substring(columns["CANCELED TIME"]!!..<columns["CANCEL FINISHED TIME"]!!).trim(),
+            cancelFinishedTime = row.substring(columns["CANCEL FINISHED TIME"]!!..<columns["CANCEL FAILED"]!!).trim(),
+            cancelFailed = row.substring(columns["CANCEL FAILED"]!!).trim(),
+        )
+    }
+}