+++ CHANGELOG.md
@@ -16,6 +16,7 @@
 - Show cloud icon for CCv2 subscription [#1051](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/1051)
 - Fetch and show CCv2 environments details [#1055](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/1055)
 - Fetch and show CCv2 builds details [#1057](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/1057)
+- Fetch and show CCv2 deployments details [#1064](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/1064)
 - Added timeout configuration for SAP CCM [#1056](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/1056)
 - Improved visual representation of the environments and builds [#1059](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/1059)
 - Open CCv2 settings pane from the toolbar [#1063](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/1063)